global variables
git-svn-id: https://yap.svn.sf.net/svnroot/yap/trunk@1685 b08c6af1-5177-4d33-ba66-4b1c6b8b522a

+++ C/tracer.c
@@ -359,7 +359,7 @@ volatile int vsc_wait;
 
 static Int p_vsc_wait(void)
 {
-  fprintf(stderr,"attach %d\n",getpid());
+  fprintf(stderr,"attach %d\n",(int)getpid());
   while (!vsc_wait);
   vsc_wait=1;
   return(TRUE);
